[common/meta] refactor: move UpsertKVAction to meta/types
