+++ common/meta/types/src/kv_message.rs
@@ -13,9 +13,20 @@
 // limitations under the License.
 
 use crate::Change;
+use crate::KVMeta;
+use crate::MatchSeq;
+use crate::Operation;
 use crate::SeqV;
 
 pub type UpsertKVActionReply = Change<Vec<u8>>;
 pub type GetKVActionReply = Option<SeqV<Vec<u8>>>;
 pub type MGetKVActionReply = Vec<Option<SeqV<Vec<u8>>>>;
 pub type PrefixListReply = Vec<(String, SeqV<Vec<u8>>)>;
+
+#[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
+pub struct UpsertKVAction {
+    pub key: String,
+    pub seq: MatchSeq,
+    pub value: Operation<Vec<u8>>,
+    pub value_meta: Option<KVMeta>,
+}
