+++ common/meta/types/src/lib.rs
@@ -33,10 +33,11 @@ pub use commit_table_reply::UpsertTableOptionReply;
 pub use database_info::DatabaseInfo;
 pub use database_reply::CreateDatabaseReply;
 pub use errors::ConflictSeq;
-pub use kv_reply::GetKVActionReply;
-pub use kv_reply::MGetKVActionReply;
-pub use kv_reply::PrefixListReply;
-pub use kv_reply::UpsertKVActionReply;
+pub use kv_message::GetKVActionReply;
+pub use kv_message::MGetKVActionReply;
+pub use kv_message::PrefixListReply;
+pub use kv_message::UpsertKVAction;
+pub use kv_message::UpsertKVActionReply;
 pub use log_entry::LogEntry;
 pub use match_seq::MatchSeq;
 pub use match_seq::MatchSeqExt;
@@ -68,7 +69,7 @@ mod commit_table_reply;
 mod database_info;
 mod database_reply;
 mod errors;
-mod kv_reply;
+mod kv_message;
 mod log_entry;
 mod match_seq;
 mod operation;
